requirements.txt
PyQt6>=6.5.0
PyCryptodome>=3.19.0
cryptography>=41.0.0
gmssl>=3.2.1
requests>=2.31.0
aiohttp>=3.8.0
